Add files via upload
animals sort have been updated. the other function used to use uniq and sort in order to determin the "other" files, but now it's a grep -v function and a counter instead. before if soemthing is not from teh array for example a file with two iron dogs the sort and uniq would eleminate it and erase it from the other option. now it wont.